requirements.txt
flask
mysql-connector-python
secure-smtplib